@@ -23,20 +23,20 @@ handlers:
       Access-Control-Allow-Origin: gs://ysfintech-homepage.appspot.com
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
 
 dependencies:
   responsive_framework: ^0.1.0
   animations: ^2.0.0
   google_fonts: ^2.0.0
   firebase_core: "^1.0.3"
-  cloud_firestore: ^1.0.4
-  firebase_storage: ^8.0.3
+  cloud_firestore: ^3.1.10
+  firebase_storage: ^10.2.9
   flutter_markdown: ^0.6.1
   url_launcher: ^6.0.3
   image_pixels: ^2.0.0
   path_provider: ^2.0.1
-  flutter_svg: ^0.19.3
+  flutter_svg: ^1.0.3
 
   flutter:
     sdk: flutter
@@ -46,8 +46,8 @@ dependencies:
   cupertino_icons: ^1.0.2
 
 dev_dependencies:
-  build_runner: ">=1.6.2 <2.0.0"
-  build_web_compilers: ">=2.12.0 <3.0.0"
+  build_runner: ^2.1.7
+  build_web_compilers: ^3.2.2
 
   flutter_test:
     sdk: flutter
