@@ -6,7 +6,7 @@ import 'package:yonsei_financial_tech/components/typography.dart';
 class TextBody extends StatelessWidget {
   final String text;
 
-  const TextBody({Key key, this.text}) : super(key: key);
+  const TextBody({Key? key, required this.text}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -23,7 +23,7 @@ class TextBody extends StatelessWidget {
 class TextBodySecondary extends StatelessWidget {
   final String text;
 
-  const TextBodySecondary({Key key, this.text}) : super(key: key);
+  const TextBodySecondary({Key? key, required this.text}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -40,7 +40,7 @@ class TextBodySecondary extends StatelessWidget {
 class TextHeadlineSecondary extends StatelessWidget {
   final String text;
 
-  const TextHeadlineSecondary({Key key, this.text}) : super(key: key);
+  const TextHeadlineSecondary({Key? key, required this.text}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -57,7 +57,7 @@ class TextHeadlineSecondary extends StatelessWidget {
 class TextBlockquote extends StatelessWidget {
   final String text;
 
-  const TextBlockquote({Key key, this.text}) : super(key: key);
+  const TextBlockquote({Key? key, required this.text}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -77,8 +77,13 @@ class TextBlockquote extends StatelessWidget {
   }
 }
 
-Text makeSmallTitle(final text, BuildContext context,
-    {Color color, FontWeight fontWeight, double fontSize}) {
+Text makeSmallTitle(
+  final text,
+  BuildContext context, {
+  Color? color,
+  FontWeight? fontWeight,
+  double? fontSize,
+}) {
   return Text(
     text,
     style: TextStyle(
@@ -87,6 +92,6 @@ Text makeSmallTitle(final text, BuildContext context,
         fontSize: fontSize != null && context != null
             ? fontSize
             : MediaQuery.of(context).size.width * 0.01),
-        textAlign: TextAlign.center,
+    textAlign: TextAlign.center,
   );
 }
