@@ -13,7 +13,6 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  ScrollController _controller = new ScrollController();
 
   // firebase cloud firestore\
   CollectionReference homes =
@@ -22,7 +21,9 @@ class _HomePageState extends State<HomePage> {
   // firebase storage
   Future<String> downloadURLExample(title) async {
     String downloadURL = await firebase_storage.FirebaseStorage.instance
-        .ref('gs://ysfintech-homepage.appspot.com/introduction/' + title + '.jpg')
+        .ref('gs://ysfintech-homepage.appspot.com/introduction/' +
+            title +
+            '.jpg')
         .getDownloadURL();
     return downloadURL;
   }
@@ -40,90 +41,80 @@ class _HomePageState extends State<HomePage> {
     final size = MediaQuery.of(context).size;
     return Scaffold(
       backgroundColor: Colors.white,
-      body: Stack(
+      body: ListView(
         children: <Widget>[
-          Scrollbar(
-            controller: _controller,
-            isAlwaysShown: true,
-            child: 
-          SingleChildScrollView(
-            controller: _controller,
-            child: Column(
-              children: <Widget>[
-                // MENU BAR ----------------------------------------------------------
-                MenuBar(),
-                // IMAGE BACKGROUND - NAME -------------------------------------------
-                title(context),
-                // PROJECTS  ------------------------------------------------------------
-                FutureBuilder<QuerySnapshot>(
-                  future: fetchedData,
-                  builder: (context, snapshot) {
-                    if (snapshot.hasError) {
-                      return Center(child: Text('500 - error'));
-                    } else if (!snapshot.hasData) {
-                      return Center(
-                          child: SizedBox(
-                              width: 40,
-                              height: 40,
-                              child: CircularProgressIndicator()));
-                    } else {
-                      var data = snapshot.data.docs.asMap();
-                      return Container(
-                          width: double.infinity,
-                          alignment: Alignment.center,
-                          child: ListView.builder(
-                            controller: _controller,  // same scroll controller 
-                            reverse: true,
-                            shrinkWrap: true,
-                            itemCount: data.length,
-                            itemBuilder: (BuildContext context, int index) {
-                              // content parsing
-                              List<String> parsedContent = data[index]
-                                  .data()['content']
-                                  .toString()
-                                  .split('<br>');
-                              String content() {
-                                StringBuffer sb = new StringBuffer();
-                                for (String item in parsedContent) {
-                                  sb.write(item + '\n\n');
-                                }
-                                return sb.toString();
-                              }
+          // MENU BAR ----------------------------------------------------------
+          MenuBar(),
+          // IMAGE BACKGROUND - NAME -------------------------------------------
+          title(context),
+          // PROJECTS  ------------------------------------------------------------
+          FutureBuilder<QuerySnapshot>(
+            future: fetchedData,
+            builder: (context, snapshot) {
+              if (snapshot.hasError) {
+                return Center(
+                    child: Text('500 - error'));
+              } else if (!snapshot.hasData) {
+                return Center(
+                    child: SizedBox(
+                        width: 40,
+                        height: 40,
+                        child: CircularProgressIndicator()));
+              } else {
+                var data = snapshot.data!.docs.asMap();
+                return Container(
+                    width: double.infinity,
+                    alignment: Alignment.center,
+                    child: ListView.builder(
+                      reverse: true,
+                      shrinkWrap: true,
+                      itemCount: data.length,
+                      itemBuilder: (BuildContext context, int index) {
+                        // content parsing
+                        List<String> parsedContent = data[index]!
+                            .get('content')
+                            .toString()
+                            .split('<br>');
+                        String content() {
+                          StringBuffer sb = new StringBuffer();
+                          for (String item in parsedContent) {
+                            sb.write(item + '\n\n');
+                          }
+                          return sb.toString();
+                        }
 
-                              return FutureBuilder(
-                                future: downloadURLExample(
-                                    data[index].data()['id'].toString()),
-                                builder: (context, snapshot) {
-                                  if (snapshot.hasError) {
-                                    return Text('500 - error');
-                                  } else if (!snapshot.hasData) {
-                                    return SizedBox(); // remove indicator
-                                  } else {
-                                    return ArticleB(
-                                      backgroundColor: index % 2 == 0
-                                          ? themeBlue
-                                          : Colors.white,
-                                      title: data[index].data()['title'],
-                                      content: content(),
-                                      role: data[index].data()['role'],
-                                      name: data[index].data()['name'],
-                                      image: Image.network(
-                                          snapshot.data.toString(),
-                                          width: 200,  // image in one size
-                                          fit: BoxFit.cover),
-                                    );
-                                  }
-                                },
+                        return FutureBuilder(
+                          future: downloadURLExample(
+                              data[index]?.get('id').toString() ?? ''),
+                          builder: (context, snapshot) {
+                            if (snapshot.hasError) {
+                              return Text(
+                                  '500 - error');
+                            } else if (!snapshot.hasData) {
+                              return SizedBox(); // remove indicator
+                            } else {
+                              return ArticleB(
+                                backgroundColor: index % 2 == 0
+                                    ? themeBlue
+                                    : Colors.white,
+                                title: data[index]?.get('title') ?? '',
+                                content: content(),
+                                role: data[index]?.get('role') ?? '',
+                                name: data[index]?.get('name') ?? '',
+                                image:
+                                    Image.network(snapshot.data.toString(),
+                                        width: 200, // image in one size
+                                        fit: BoxFit.cover),
                               );
-                            },
-                          ));
-                    }
-                  },
-                ),
-                Footer(),
-              ],
-            ),
-          )),
+                            }
+                          },
+                        );
+                      },
+                    ));
+              }
+            },
+          ),
+          Footer(),
         ],
       ),
     );
